operations research/requirements.txt
altgraph==0.17.4
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3
kiwisolver==1.4.8
matplotlib==3.10.0
networkx==3.4.2
numpy==2.2.1
packaging==24.1
pefile==2023.2.7
pillow==10.3.0
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.2
segno==1.6.1
setuptools==70.1.1
six==1.17.0